backend/requirements-render.txt
# VISHWAS Backend - Render Free Tier Optimized
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document Processing (Lightweight)
pymupdf==1.23.8
PyPDF2==3.0.1

# AI & APIs (Cloud-based)
openai==1.3.5
google-cloud-vision==3.4.4
httpx==0.25.2

# Web Search & APIs
googlesearch-python==1.2.3
duckduckgo-search==3.9.3
wikipedia==1.4.0

# Database & ORM
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0

# NLP (Lightweight)
spacy==3.7.2
